// src/Models/Interfaces/IEnderecoModel.ts
export interface IEnderecoModel {
    codigo: string
    codigoRef: string
    cep: string
    tipoEndereco: string
    rua: string
    numero: string
    bairro: string
    cidade: string
    cidadeID: string
    codigoIBGECidade: string
    estado: string
    estadoID: string
    codigoIBGEEstado: string
}